OCSSD based global translation test - 6
A lot of debug sequece can be served by me.
If you prepare the 200GB host OCSSD and run that
OCSSD in the virtual disk, prepare the fio 30GB, numjobs=1, io=write
then this will stuck in the 97% of the disk write(which has 529 free lines).

This commit contains a lot of chenges. Therefore, I cannot write down
the all of changes in this commit.

If you want to know the specific changes then follow 3 sequence.
First, you have to check the source codes commments.
Second, contact the main programmer(BlaCkinkGJ) of this commit.